Module filecoin_proofs::constants

Structs

LAYERS

POREP_MINIMUM_CHALLENGES

POREP_PARTITIONS

PorepMinimumChallenges

WINDOW_POST_SECTOR_COUNT

Constants

DRG_DEGREE

The base degree used for all DRG graphs. One degree from this value is used to ensure that a given node always has its immediate predecessor as a parent, thus ensuring unique topological ordering of the graph nodes.

EXP_DEGREE

The expansion degree used for Stacked Graphs.

MAX_LEGACY_REGISTERED_SEAL_PROOF_ID

MINIMUM_RESERVED_BYTES_FOR_PIECE_IN_FULLY_ALIGNED_SECTOR

MINIMUM_RESERVED_LEAVES_FOR_PIECE_IN_SECTOR

MIN_PIECE_SIZE

The minimum size a single piece must have before padding.

PUBLISHED_SECTOR_SIZES

Sector sizes for which parameters have been published.

SECTOR_SIZE_1_GIB

SECTOR_SIZE_2_KIB

SECTOR_SIZE_4_KIB

SECTOR_SIZE_8_MIB

SECTOR_SIZE_16_KIB

SECTOR_SIZE_16_MIB

SECTOR_SIZE_32_GIB

SECTOR_SIZE_32_KIB

SECTOR_SIZE_64_GIB

SECTOR_SIZE_512_MIB

SINGLE_PARTITION_PROOF_LEN

The size of a single snark proof.

TEST_SEED

WINDOW_POST_CHALLENGE_COUNT

WINNING_POST_CHALLENGE_COUNT

WINNING_POST_SECTOR_COUNT

Functions

is_sector_shape_base

is_sector_shape_sub2

is_sector_shape_sub8

is_sector_shape_top2

Type Aliases

DefaultBinaryTree

A binary merkle tree with Poseidon hashing, where all levels have arity 2. It’s fully persisted to disk.

DefaultOctLCTree

A merkle tree with Poseidon hashing, where all levels have arity 8. Some levels are not persisted to disk, but only cached in memory.

DefaultOctTree

A merkle tree with Poseidon hashing, where all levels have arity 8. It’s fully persisted to disk.

DefaultPieceDomain

DefaultPieceHasher

The hasher used for creating comm_d.

DefaultTreeDomain

DefaultTreeHasher

The default hasher for merkle trees currently in use.

SectorShape1GiB

SectorShape2KiB

SectorShape4KiB

SectorShape8MiB

SectorShape16KiB

SectorShape16MiB

SectorShape32GiB

SectorShape32KiB

SectorShape64GiB

SectorShape512MiB

SectorShapeBase

SectorShapeSub2

SectorShapeSub8

SectorShapeTop2