Function filecoin_proofs::get_seal_inputs

pub fn get_seal_inputs<Tree: 'static + MerkleTreeTrait>(
    porep_config: &PoRepConfig,
    comm_r: Commitment,
    comm_d: Commitment,
    prover_id: ProverId,
    sector_id: SectorId,
    ticket: Ticket,
    seed: Ticket
) -> Result<Vec<Vec<Scalar>>>

Given the specified arguments, this method returns the inputs that were used to generate the seal proof. This can be useful for proof aggregation, as verification requires these inputs.

This method allows them to be retrieved when needed, rather than storing them for some amount of time.

Arguments

porep_config - this sector’s porep config that contains the number of bytes in the sector.

comm_r - a commitment to a sector’s replica.

comm_d - a commitment to a sector’s data.

prover_id - the prover_id used to seal this sector.

sector_id - the sector_id of this sector.

ticket - the ticket used to generate this sector’s replica-id.

seed - the seed used to derive the porep challenges.