Module filecoin_proofs::types

Structs

EmptySectorUpdateEncoded

EmptySectorUpdateProof

FallbackPoStSectorProof

Labels

MerkleProof

Representation of a merkle proof.

PaddedBytesAmount

PartitionSnarkProof

PersistentAux

Stored along side the sector on disk.

PieceInfo

PoRepConfig

PoRepProofBytesAmount

PoRepProofPartitions

PoStConfig

PoStProofBytesAmount

PoStProofPartitions

PrivateReplicaInfo

The minimal information required about a replica, in order to be able to generate a PoSt over it.

PublicReplicaInfo

The minimal information required about a replica, in order to be able to verify a PoSt over it.

SealCommitOutput

SealCommitPhase1Output

SealPreCommitOutput

SealPreCommitPhase1Output

SectorClass

SectorSize

SectorUpdateConfig

StoreConfig

TemporaryAux

UnpaddedByteIndex

UnpaddedBytesAmount

UpdateProofPartitions

Enums

PoStType

Constants

BINARY_ARITY

Arity for binary trees, used for comm_d.

OCT_ARITY

Arity for oct trees, used for comm_r_last.

Traits

MerkleTreeTrait

Trait used to abstract over the way Merkle Trees are constructed and stored.

Type Definitions

AggregateSnarkProof

ChallengeSeed

Commitment

DataTree

PartitionProof

ProverId

SnarkProof

Ticket

VanillaProof

VanillaSealProof