Module filecoin_proofs::caches

Structs

SRSCache

We have a separate SRSCache type for srs keys since they are cached differently (as a hashmap per type, keyed by identifier consisting of sector size and pow2 num proofs to aggregate).

Functions

cache_lookup

get_empty_sector_update_params

get_empty_sector_update_verifying_key

get_post_verifying_key

get_stacked_params

get_stacked_srs_key

get_stacked_srs_verifier_key

get_stacked_verifying_key

lookup_groth_params

lookup_srs_key

lookup_srs_verifier_key

lookup_verifying_key

srs_cache_lookup

Type Aliases

Bls12PreparedVerifyingKey