Function filecoin_proofs::generate_tree_r_last

pub fn generate_tree_r_last<O, R, TreeR: 'static + MerkleTreeTrait>(
    sector_size: u64,
    replica_path: R,
    output_dir: O
) -> Result<<TreeR::Hasher as Hasher>::Domain>
where
    O: AsRef<Path>,
    R: AsRef<Path>,

Generate the merkle tree on top of the replica (TreeRLast).

The generated trees are stored in output_dir, usually the cache directory. The replica_path point to the replica where the tree should be built upon. The sector_size is in bytes.